abstract away concrete implementation of data storage
The goal of these changes is to be able to use BigInts
instead of the mixed Int/Long/BigInt representation
through a command line option.
To this end we are working on ensuring that only the
specialized DataStore and ExpressionCompiler depend on the
underlying value representation.
This work will eventually enable us to compute on symbolic values.